sync: DZ pre 20231013  * removed OpenZwave  * added Alfen Eve charge station support  * added Enever.nl dynamic energy pricing info support  * added Mitsubishi WF RAC Airco support

+++ src/hardware/KMTronicTCP.cpp
@@ -3,7 +3,6 @@
 #include "main/Logger.h"
 #include "main/Helper.h"
 #include <iostream>
-#include "main/localtime_r.h"
 #include "main/mainworker.h"
 #include "protocols/HTTPClient.h"
 #include "protocols/UrlEncode.h"
@@ -227,7 +226,7 @@ void KMTronicTCP::ParseRelays(const std::string &sResult)
 			{
 				bool bIsOn = (tmpstr[jj] != '0');
 				std::stringstream sstr;
-				int iRelay = (jj + 1);
+				int iRelay = static_cast<int>(jj + 1);
 				sstr << "Relay " << iRelay;
 				SendSwitch(iRelay, 1, 255, bIsOn, 0, sstr.str(), m_Name);
 				if (iRelay > m_TotRelais)
