+++ src/hardware/Limitless.cpp
@@ -4,7 +4,6 @@
 #include "main/Logger.h"
 #include "main/SQLHelper.h"
 #include "hardware/hardwaretypes.h"
-#include "main/localtime_r.h"
 #include "main/mainworker.h"
 #include "main/WebServer.h"
 #include "webserver/cWebem.h"
@@ -283,7 +282,7 @@ bool CLimitLess::IsDataAvailable(const SOCKET /*sock*/)
 	tv.tv_usec = 0;
 
 	// Wait until timeout or data received.
-	n = select(m_RemoteSocket + 1, &fds, nullptr, nullptr, &tv);
+	n = select(static_cast<int>(m_RemoteSocket + 1), &fds, nullptr, nullptr, &tv);
 	return (n > 0);
 }
