+++ src/hardware/YouLess.cpp
@@ -5,7 +5,6 @@
 #include "protocols/HTTPClient.h"
 #include "protocols/UrlEncode.h"
 #include "hardware/hardwaretypes.h"
-#include "main/localtime_r.h"
 #include "main/mainworker.h"
 #include "main/json_helper.h"
 
@@ -203,7 +202,7 @@ void CYouLess::GetMeterDetails()
 		Log(LOG_ERROR,"Error connecting to: %s", m_szIPAddress.c_str());
 		return;
 	}
-	int fpos;
+	size_t fpos;
 	std::string pusage=stdstring_trim(results[0]);
 	fpos = pusage.find_first_of(' ');
 	if (fpos!=std::string::npos)
